+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Introduced a new `perVisit()` function for the Sustainable Web Design model, which applies [caching and return visits assumptions](https://sustainablewebdesign.org/calculating-digital-emissions/).
+
+## 0.10.2 2022-08-12
+
 - Added the ability to set the model used by CO2.js to the Sustainable Web Design model, using a simple 'swd' string, instead of needing to pass in a class.
 
 ## 0.10.1 2022-08-01
@@ -26,7 +30,7 @@ This release used a version bump as previously we had released v0.10.0 under a p
 
 - Added ES import syntax as the main way for handling imports and exports of code within the module.
 - Changed eslint settings to use later version of ecmascript (2020)
-- Change the bulid tools to use esbulid with jest instead of babel
+- Change the build tools to use esbulid with jest instead of babel
 - Added more consistent use of the debug logging library in files using the updated import syntax
 - Fixed the incorrect order of FIRST_TIME_VIEWING_PERCENTAGE and RETURNING_VISITOR_PERCENTAGE constants in the SWD model. This will result in **larger** values for calculations using the sustainable web design, and the default caching assumptions.